+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # [Unreleased]
 
+## [v1.0.5] - 2021-09-08
+- [Issue #12](https://github.com/marcodallasanta/ssh-scp-deploy/issues/12) - Action never failed also when file/dir was not uploaded.
+
 ## [v1.0.4] - 2021-02-10
 
 ### Fixed
-- Issue #9 - Wrong if on password.
+- [Issue #9](https://github.com/marcodallasanta/ssh-scp-deploy/issues/9) - Wrong if on password.
 
 ## [v1.0.3] - 2021-02-08
 
@@ -27,6 +30,7 @@
 
 
 [unreleased]: https://github.com/marcodallasanta/ssh-scp-deploy/compare/main...development
+[v1.0.5]: https://github.com/underscore69/ssh-scp-deploy/tree/v1.0.5
 [v1.0.4]: https://github.com/underscore69/ssh-scp-deploy/tree/v1.0.4
 [v1.0.3]: https://github.com/underscore69/ssh-scp-deploy/tree/v1.0.3
 [v1.0.2]: https://github.com/underscore69/ssh-scp-deploy/tree/v1.0.2